Deletes separate user routes
