+++ app.js
@@ -4,10 +4,7 @@ const cors = require('cors')
 require('dotenv').config()
 
 const contactsRouter = require('./routes/api/contacts');
-const registerRouter = require('./routes/api/register');
-const loginRouter = require('./routes/api/login');
-const logoutRouter = require('./routes/api/logout');
-const currentRouter = require('./routes/api/current');
+const usersRouter = require('./routes/api/users');
 
 const app = express()
 
@@ -18,10 +15,7 @@ app.use(cors())
 app.use(express.json())
 
 app.use('/api/contacts', contactsRouter);
-app.use('/api/users/register', registerRouter);
-app.use('/api/users/login', loginRouter);
-app.use('/api/users/logout', logoutRouter);
-app.use('/api/users/current', currentRouter);
+app.use('/api/users', usersRouter);
 
 app.use((req, res) => {
   res.status(404).json({ message: 'Not found' })
