--- routes/api/logout.js
@@ -1,10 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const { ctrlWrapper } = require('../../helpers/index');
-const { logout } = require('../../controllers/auth/index');
-const authenticate = require('../../middlewares/authenticate');
-
-
-router.get("/", authenticate, ctrlWrapper(logout));
-
-module.exports = router;
